resturcutre audioplayer ui code
