+++ new-player/src/main/java/net/newpipe/newplayer/ui/audioplayer/AudioPlayerUI.kt
@@ -171,231 +171,6 @@ internal fun AudioPlayerUI(
 }
 
 @OptIn(UnstableApi::class)
-@Composable
-private fun LandscapeLayout(
-    viewModel: InternalNewPlayerViewModel,
-    uiState: NewPlayerUIState,
-    modifier: Modifier = Modifier,
-) {
-    Column(modifier = modifier) {
-
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(32.dp),
-        ) {
-
-            LandscapeCoverArtOrPlaylistUI(
-                modifier = Modifier.weight(1f),
-                viewModel = viewModel,
-                uiState = uiState,
-            )
-
-            Column(
-                verticalArrangement = Arrangement.SpaceEvenly,
-                modifier = Modifier
-                    .fillMaxHeight()
-                    .weight(1f)
-            ) {
-                AnimatedVisibility(uiState.showPlaylistInAudioPlayer) {
-                    StreamSelectTopBar(
-                        viewModel = viewModel,
-                        uiState = uiState,
-                        isUsedInAudiPlayer = true
-                    )
-                }
-
-                AudioPlaybackController(viewModel = viewModel, uiState = uiState)
-                ProgressUI(viewModel = viewModel, uiState = uiState)
-                AudioBottomUI(viewModel = viewModel, uiState = uiState)
-            }
-        }
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun PortraitLayout(
-    viewModel: InternalNewPlayerViewModel,
-    uiState: NewPlayerUIState,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = Modifier,
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        androidx.compose.animation.AnimatedVisibility(uiState.showPlaylistInAudioPlayer) {
-            StreamSelectTopBar(viewModel = viewModel, uiState = uiState, isUsedInAudiPlayer = true)
-        }
-        Column(
-            modifier = modifier,
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(32.dp),
-        ) {
-            PortraitCoverArtOrPlaylistUI(
-                modifier = Modifier
-                    .fillMaxHeight()
-                    .weight(1f),
-                viewModel = viewModel,
-                uiState = uiState
-            )
-            AudioPlaybackController(viewModel = viewModel, uiState = uiState)
-            ProgressUI(viewModel = viewModel, uiState = uiState)
-            AudioBottomUI(viewModel, uiState)
-        }
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun PortraitCoverArtOrPlaylistUI(
-    modifier: Modifier = Modifier,
-    viewModel: InternalNewPlayerViewModel,
-    uiState: NewPlayerUIState
-) {
-    Box(modifier = modifier.fillMaxSize()) {
-        AnimatedVisibility(uiState.showPlaylistInAudioPlayer, enter = fadeIn(), exit = fadeOut()) {
-            ReorderableStreamItemsList(
-                modifier = Modifier
-                    .fillMaxSize(),
-                viewModel = viewModel,
-                uiState = uiState
-            )
-        }
-
-        AnimatedVisibility(!uiState.showPlaylistInAudioPlayer, enter = fadeIn(), exit = fadeOut()) {
-            Column(
-                modifier = Modifier.fillMaxSize(),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center,
-            ) {
-                CoverArt(uiState = uiState)
-                Spacer(modifier = Modifier.height(24.dp))
-                TitleView(uiState = uiState)
-            }
-        }
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun LandscapeCoverArtOrPlaylistUI(
-    modifier: Modifier = Modifier,
-    viewModel: InternalNewPlayerViewModel,
-    uiState: NewPlayerUIState
-) {
-    Box(modifier = modifier.fillMaxSize()) {
-        AnimatedVisibility(uiState.showPlaylistInAudioPlayer, enter = fadeIn(), exit = fadeOut()) {
-            ReorderableStreamItemsList(
-                modifier = modifier
-                    .fillMaxSize(),
-                viewModel = viewModel,
-                uiState = uiState
-            )
-        }
-        AnimatedVisibility(!uiState.showPlaylistInAudioPlayer, enter = fadeIn(), exit = fadeOut()) {
-            Column(modifier = modifier) {
-                TitleView(
-                    modifier = Modifier
-                        .fillMaxWidth(),
-                    uiState = uiState,
-                )
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                CoverArt(
-                    uiState = uiState
-                )
-            }
-        }
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun ProgressUI(
-    modifier: Modifier = Modifier,
-    viewModel: InternalNewPlayerViewModel,
-    uiState: NewPlayerUIState
-) {
-    val locale = getLocale()!!
-
-    Column(modifier = modifier) {
-        Box(
-            modifier = Modifier
-                .height(0.dp)
-                .fillMaxWidth()
-                .wrapContentHeight(unbounded = true, align = Alignment.Bottom)
-        ) {
-            ThumbPreview(
-                modifier = Modifier.offset(y = (-20).dp) /* We have this offset to make space for your thumb */,
-                uiState = uiState,
-                thumbSize = SeekerDefaults.ThumbRadius * 2,
-                previewHeight = 120.dp
-            )
-        }
-
-        NewPlayerSeeker(viewModel = viewModel, uiState = uiState)
-        Row {
-            Text(
-                getTimeStringFromMs(
-                    uiState.playbackPositionInMs,
-                    getLocale() ?: locale
-                )
-            )
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .weight(1f)
-            )
-            Text(
-                getTimeStringFromMs(
-                    uiState.durationInMs,
-                    getLocale() ?: locale
-                )
-            )
-        }
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun TitleView(modifier: Modifier = Modifier, uiState: NewPlayerUIState) {
-    Column(modifier = modifier) {
-        Text(
-            text = uiState.currentlyPlaying?.mediaMetadata?.title.toString(),
-            maxLines = 1,
-            style = MaterialTheme.typography.headlineMedium,
-            modifier = Modifier.basicMarquee(),
-        )
-        Spacer(modifier = Modifier.height(6.dp))
-        Text(
-            text = uiState.currentlyPlaying?.mediaMetadata?.artist.toString(),
-            maxLines = 1,
-            style = MaterialTheme.typography.bodyLarge,
-            modifier = Modifier.basicMarquee(),
-        )
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun CoverArt(modifier: Modifier = Modifier, uiState: NewPlayerUIState) {
-    Box(modifier = modifier) {
-        Card(
-            elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
-        ) {
-            Thumbnail(
-                modifier = Modifier.fillMaxWidth(),
-                thumbnail = uiState.currentlyPlaying?.mediaMetadata?.artworkUri,
-                contentDescription = stringResource(
-                    id = R.string.stream_thumbnail
-                ),
-            )
-        }
-    }
-}
-
-@OptIn(UnstableApi::class)
 @Preview(device = "id:pixel_6", showSystemUi = true)
 @Composable
 private fun AudioPlayerUIPortraitPreview() {
